+++ detail/client/src/main/java/au/com/deanpike/detail/client/usecase/ProjectDetailUseCase.kt
@@ -4,5 +4,5 @@ import au.com.deanpike.commonshared.util.ResponseWrapper
 import au.com.deanpike.detail.client.model.detail.ProjectDetail
 
 interface ProjectDetailUseCase {
-    suspend fun getProjectDetails(id: Int): ResponseWrapper<ProjectDetail>
+    suspend fun getProjectDetails(id: Long): ResponseWrapper<ProjectDetail>
 }
